+++ src/main/java/jupiter/extension/ProjectExtension.java
@@ -3,7 +3,9 @@ package jupiter.extension;
 import api.models.CreateProjectRequest;
 import api.models.CreateUserResponse;
 import api.models.ParentProject;
+import api.models.ProjectResponse;
 import api.requests.steps.ProjectManagementSteps;
+import common.data.ProjectData;
 import common.generators.TestDataGenerator;
 import jupiter.annotation.WithProject;
 import org.junit.jupiter.api.extension.*;
@@ -12,38 +14,54 @@ public class ProjectExtension implements BeforeEachCallback, ParameterResolver {
 
     public static final ExtensionContext.Namespace NAMESPACE =
             ExtensionContext.Namespace.create(ProjectExtension.class);
-    private CreateUserResponse user;
 
     @Override
     public void beforeEach(ExtensionContext context) throws Exception {
         WithProject annotation = context.getRequiredTestMethod().getAnnotation(WithProject.class);
 
         if (annotation != null) {
-            user = context.getStore(UsersQueueExtension.NAMESPACE)
+            CreateUserResponse user = context.getStore(UsersQueueExtension.NAMESPACE)
                     .get(context.getUniqueId(), CreateUserResponse.class);
 
             if (user == null) {
                 throw new ExtensionConfigurationException("Добавьте аннотацию @WithUsersQueue");
             }
 
-            CreateProjectRequest project = CreateProjectRequest.builder()
-                    .id(TestDataGenerator.generateProjectID())
-                    .name(TestDataGenerator.generateProjectName())
-                    .parentProject(new ParentProject(annotation.parentProjectId().value))
-                    .build();
+            String parentProjectId = annotation.parentProjectId().equals("default") ? ProjectData.PARENT_PROJECT.value : annotation.parentProjectId();
+            String projectId = annotation.projectId().equals("default") ? TestDataGenerator.generateProjectID() : annotation.projectId();
+            String projectName = annotation.projectName().equals("default") ? TestDataGenerator.generateProjectName() : annotation.projectName();
 
-            ProjectManagementSteps.createProject(
-                    project.getId(),
-                    project.getName(),
-                    project.getParentProject(),
-                    user
-            );
+            CreateProjectRequest project;
+
+            if(annotation.useExisting()) {
+                ProjectResponse projectById = ProjectManagementSteps.getProjectById(projectId, user);
+                project = CreateProjectRequest.builder()
+                        .parentProject(new ParentProject(projectById.getParentProject().getId()))
+                        .id(projectById.getId())
+                        .name(projectById.getName())
+                        .build();
+            } else {
+                ProjectResponse projectById = ProjectManagementSteps.createProject(
+                        projectId,
+                        projectName,
+                        new ParentProject(parentProjectId),
+                        user
+                );
+
+                project = CreateProjectRequest.builder()
+                        .parentProject(new ParentProject(projectById.getParentProject().getId()))
+                        .id(projectById.getId())
+                        .name(projectById.getName())
+                        .build();
+            }
 
             context.getStore(NAMESPACE).put(context.getUniqueId(), project);
 
-            context.getStore(NAMESPACE).put("project_cleanup", (ExtensionContext.Store.CloseableResource) () -> {
+            if(annotation.addToCleanup() && !annotation.useExisting()) {
+                context.getStore(NAMESPACE).put("project_cleanup", (ExtensionContext.Store.CloseableResource) () -> {
                     ProjectManagementSteps.deleteProjectByIdQuietly(project.getId(), user);
-            });
+                });
+            }
         }
     }
